Rename qui marche de partout (+ pour les décorateurs)
